play with how the mesh is generated.
Mesh generation was working how I liked.

Attempted to go with a different approach for when I randomly generate things with poisson disc sampling.

Consider raising mesh after placing items. But then I remember the only thing actually keeping the objects on the mesh was the collider.

Maybe there is a way to get an object place on the map, and then just put it at the new height.

Maybe it's something similar to onCollision, and then you can get that point on the mesh where it is colliding?

I'll look into it later.